Collect a listing of the changes made and add to the commit message
I am dropping the file in `.git/` because the scripts run within the context of the repository the Action works on, not its own checkout, so we can't use our `.gitignore`.

I didn't want to muck around with getting a temporary directory path in a GitHub Actions compatible way or figuring out how to read the file to use later (as a step output?) and deleting it before staging the index, so I decided to just shelve it to the `.git` directory where Git won't go looking for changes.